+++ tests/test_bwaise.py
@@ -16,6 +16,6 @@ def test_bwaise():
     from exposan import bwaise as bw
     bw.print_summaries((bw.sysA, bw.sysB, bw.sysC))
     
-# This just means that if pytest runs this module, it calls the test_bwaise function
+# If pytest runs this module, it calls the test_bwaise function and test exposan
 if __name__ == '__main__':
     test_bwaise()